Refactor: move dump import/export code to stores; ensure caches are invalidated on dump import (#156)
